requirements.txt
jira==2.0.0
requests==2.19.0
PyMySQL==0.9.3
xmltodict==0.11.0
tornado==5.1
gitlab==1.0.2
python-gitlab==1.8.0
jenkins==1.0.2
PyYAML==5.1